requirements.txt
google-cloud-bigquery==1.26.0
google-cloud-core==1.3.0
google-cloud-datacatalog==1.0.0
pyyaml
